Added a -r (reallyclean) switch to control if bin, lib, etc., are removed.

+++ hyrax_clean.sh
@@ -1,6 +1,32 @@
 #!/bin/sh
 #
-# Clean it all; not very elegant
+# Clean it all; not very elegant; -r == 'really clean' removes bin, lib, ...
+
+args=`getopt r $*`
+if [ $? != 0 ]
+then
+    echo "Usage: $0 [options] where options are:"
+    echo "-r: really clean; remove bin, libs, ..., too"
+    exit 2
+fi
+
+set -- $args
+
+# Set verbose and do_nothing to false
+vernose=""
+realclean=""
+for i in $*
+do
+    case "$i"
+	in
+        -r)
+            reallyclean="yes";
+            shift;;
+        --)
+            shift; break;;
+    esac
+done
+
 
 (cd libdap && make clean)
 
@@ -14,5 +40,7 @@ then
 fi
 ) # 'cd modules sub-shell
 
-rm -rf bin etc include lib share var
-
+if test -n "$reallyclean"
+then
+    rm -rf bin etc include lib share var
+fi
